# tools/openapi.yaml
openapi: 3.0.0
info:
  title: Parking Fee Calculation API
  version: 0.1.0
servers:
  - url: /
paths:
...
  /api/rate_plans/load/{plan_id}:
    get:
      summary: 讀取自訂方案
      parameters:
        - in: path
          name: plan_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
  /api/rate_plans/save:
    post:
      summary: 儲存自訂方案
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: 成功
  /api/rate_plans/{plan_id}:
    delete:
      summary: 刪除自訂方案
      parameters:
        - in: path
          name: plan_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
...
  /api/mdp/templates/{template_id}:
    get:
      summary: 讀取模板
      parameters:
        - in: path
          name: template_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
    delete:
      summary: 刪除模板
      parameters:
        - in: path
          name: template_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
  /api/mdp/templates/save:
    post:
      summary: 儲存模板
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: 成功
...
  /api/mdp/preview:
    post:
      summary: 快速預覽試算
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: 成功
  /api/enhanced/segments/validate:
    post:
      summary: 驗證 24 小時區段覆蓋與重疊
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: 成功
...